refactor edit customer component
