+++ carRental/src/pages/Edit/EditCustomer/EditCustomerPage.jsx
@@ -11,13 +11,15 @@ import {
 } from "../../../utils/constants/general-labels";
 import PageTitle from "../../../components/PageTitle/PageTitle";
 import "./EditCustomerPage.css";
-import { updateCustomer } from "../../../utils/parse-functions/updateFunctions";
+import {
+  updateCustomer,
+  getCustomerById,
+} from "../../../utils/parse-functions/updateFunctions";
 import {
   onChangeIntHandler,
   onChangeHandler,
 } from "../../../utils/functions/onChangeHandlers";
 import PopUpButton from "../../../components/PopUpButton/PopUpButton";
-import Parse from "parse";
 
 const EditCustomerPage = () => {
   const { bookingId, customerId } = useParams();
@@ -25,31 +27,8 @@ const EditCustomerPage = () => {
   const [formData, setFormData] = useState();
   const [customer, setCustomer] = useState([]);
 
-  const getCustomerById = async () => {
-    const Customer = Parse.Object.extend("Customer");
-    const query = new Parse.Query(Customer);
-    query.equalTo("objectId", customerId);
-    try {
-      const results = await query.find();
-      for (const object of results) {
-        const customerObject = {
-          lastName: object.get("lastName"),
-          driversLicenseID: object.get("driversLicenseID"),
-          phoneNumber: object.get("phoneNumber"),
-          firstName: object.get("firstName"),
-          email: object.get("email"),
-          age: object.get("age"),
-          address: object.get("address"),
-        };
-        setCustomer(customerObject);
-      }
-    } catch (error) {
-      console.error("Error while fetching Customer", error);
-    }
-  };
-
   useEffect(async () => {
-    await getCustomerById();
+    await getCustomerById(customerId, setCustomer);
   }, []);
   console.log(customer);
